Move deepCheck to utils
